pyproject.toml
[tool.poetry]
name = "spotify-e-paper-control"
version = "0.1.0"
description = "Remote control Spotify via touchscreen e-paper display"
authors = ["Cutwell"]
license = "MIT"
readme = "README.md"
packages = [{include = "spotify_e_paper_control"}]

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.23.0"
segno = "^1.6.0"

[tool.poetry.group.rpi.dependencies]
pillow = "^10.2.0"
rpi-gpio = "^0.7.1"
gpiozero = "^2.0"
smbus = "^1.1.post2"
spidev = "^3.6"
lgpio = "^0.0.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
